# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin框架的快速开发脚手架API文档",
        "title": "Gin-Fast API",
        "termsOfService": "https://github.com/your-repo/gin-fast",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/your-repo/gin-fast/issues",
            "email": "your-email@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/your-repo/gin-fast/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/sysUserTenant/batchAdd": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "批量创建用户租户关联",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "批量新增用户租户关联",
                "parameters": [
                    {
                        "description": "批量用户租户关联信息",
                        "name": "sysUserTenant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SysUserTenantBatchAddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户租户关联批量创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sysUserTenant/batchDelete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "批量删除指定用户租户关联",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "批量删除用户租户关联",
                "parameters": [
                    {
                        "description": "批量用户租户关联删除请求参数",
                        "name": "sysUserTenant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SysUserTenantBatchDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户租户关联批量删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sysUserTenant/get": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据用户ID和租户ID获取用户租户关联详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "根据用户ID和租户ID获取用户租户关联信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "租户ID",
                        "name": "tenantID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回用户租户关联信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数格式错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sysUserTenant/getRolesAll": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取所有角色列表，支持分页和过滤，不进行租户过滤，可根据tenant_id查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "角色列表(不进行租户过滤)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "父级ID",
                        "name": "parentId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回角色列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sysUserTenant/getUserRoleIDs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据tenant_id和user_id查询角色ID集合，注意sys_user_role表没有tenant_id字段，sys_role表中才有tenant_id字段",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "根据tenant_id和user_id查询角色ID集合",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回角色ID集合",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数格式错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sysUserTenant/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户租户关联列表，支持分页和过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "用户租户关联列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "userID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "租户ID",
                        "name": "tenantID",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回用户租户关联列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sysUserTenant/setUserRoles": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "先移除租户相关的角色，再为用户分配新角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "设置用户角色",
                "parameters": [
                    {
                        "description": "设置用户角色请求参数",
                        "name": "setUserRoles",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SysUserTenantSetRolesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置用户角色成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/sysUserTenant/userListAll": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取所有用户列表，支持分页和过滤，不进行租户过滤",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户租户关联管理"
                ],
                "summary": "用户列表(不进行租户过滤)",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户名或昵称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回用户列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.SysUserTenantBatchAddRequest": {
            "type": "object",
            "required": [
                "tenantID",
                "userIDs"
            ],
            "properties": {
                "tenantID": {
                    "type": "integer"
                },
                "userIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.SysUserTenantBatchDeleteRequest": {
            "type": "object",
            "required": [
                "tenantID",
                "userIDs"
            ],
            "properties": {
                "tenantID": {
                    "type": "integer"
                },
                "userIDs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.SysUserTenantSetRolesRequest": {
            "type": "object",
            "required": [
                "roles",
                "tenantID",
                "userID"
            ],
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tenantID": {
                    "type": "integer"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
    }
}